feat: Implement major content expansion and i18n for JB SQUARE
This commit delivers a major feature expansion for the JB SQUARE project, turning it into a comprehensive and interactive single-page application.

Key Changes:

1.  **Component Refactoring:** The entire application, previously in a single file, has been refactored into a clean, maintainable structure with separate component files in `src/components`.

2.  **Internationalization (i18n):**
    - Implemented full bilingual (Korean/English) support using `react-i18next`.
    - Added a functional language toggle button in the navigation.
    - All user-facing text is now managed through locale JSON files.

3.  **Content & Feature Expansion:**
    - Replaced broken image links in the Companies section with reliable placeholders.
    - Enhanced the Footer with a newsletter subscription form and social media links.
    - All mock data arrays have been expanded by ~3x to create a more realistic and complete user experience.

This commit addresses all user requirements for this development phase, resulting in a robust, maintainable, and feature-rich application.
