+++ README.md
@@ -6,15 +6,12 @@ The page features a rich, multi-section layout with detailed mock data, dropdown
 
 ## Key Features & Sections
 
--   **Dynamic Navigation:** A fully functional dropdown navigation menu.
--   **Hero Section:** An impactful introduction to JB SQUARE.
--   **Stats:** Key metrics about the bio-cluster.
--   **Announcements:** A list of the latest opportunities and notices (expanded data).
--   **Events:** A grid of upcoming events (expanded data).
--   **Companies:** A directory of breakthrough companies (expanded data).
--   **Support Ecosystem:** A new section showcasing partner organizations.
--   **Policy & Investment:** A new tabbed section detailing policies and funds.
--   **Research & Patents:** A new section listing recent IP and publications.
+-   **Component-Based Architecture:** All UI sections are refactored into their own separate components for better maintainability.
+-   **Internationalization (i18n):** Fully bilingual support for English and Korean, implemented with `react-i18next`.
+-   **Dynamic Navigation:** A functional dropdown navigation menu that is also translated.
+-   **Expanded Content:** All sections now feature a larger, more realistic set of mock data.
+-   **New Sections:** Includes newly added sections for Support Ecosystem, Policy & Investment, and Research & Patents.
+-   **Enhanced Footer:** The footer is now more functional with a newsletter form and social media links.
 
 ## Technologies Used
 
@@ -22,6 +19,7 @@ The page features a rich, multi-section layout with detailed mock data, dropdown
 -   **Language:** TypeScript
 -   **Build Tool:** Vite
 -   **Styling:** Tailwind CSS (v3 - Stable)
+-   **Internationalization:** `react-i18next` & `i18next`
 -   **Animations:** Framer Motion
 -   **Icons:** Lucide React
 
@@ -43,4 +41,4 @@ After installation, run the following command:
 npm run dev
 ```
 
-This will start the development server, typically on `http://localhost:5173`. Open this URL in your browser to view the live application.
+This will start the development server, typically on `http://localhost:5173`. Open this URL in your browser to view the live application. You can switch between English and Korean using the language toggle in the header.
