+++ package.json
@@ -11,9 +11,12 @@
   },
   "dependencies": {
     "framer-motion": "^12.23.12",
+    "i18next": "^23.11.5",
+    "i18next-http-backend": "^2.5.2",
     "lucide-react": "^0.542.0",
     "react": "^19.1.1",
-    "react-dom": "^19.1.1"
+    "react-dom": "^19.1.1",
+    "react-i18next": "^14.1.2"
   },
   "devDependencies": {
     "@eslint/js": "^9.33.0",
