+++ src/components/EventsSection.tsx
@@ -1,0 +1,188 @@
+import React from 'react';
+import { motion } from 'framer-motion';
+import { Calendar, MapPin } from 'lucide-react';
+import { useTranslation } from 'react-i18next';
+import { SectionProps } from './types';
+import { boldFadeIn, staggerBold } from './animations';
+
+const EventsSection = ({ id }: SectionProps) => {
+  const { t } = useTranslation();
+
+  const events = [
+    {
+      id: 1,
+      title: "Jeonbuk BioTech Innovation Summit 2025",
+      date: "October 15, 2025",
+      location: "JBFEZ Convention Center",
+      featured: true,
+      category: "SUMMIT",
+      image: "https://images.unsplash.com/photo-1511578314322-379afb476865?ixlib=rb-4.0.3&auto=format&fit=crop&w=800&q=80",
+      attendees: "500+"
+    },
+    {
+      id: 2,
+      title: "CEO Forum: Digital Health Revolution",
+      date: "September 28, 2025",
+      location: "Bio Valley Complex",
+      featured: false,
+      category: "FORUM",
+      image: "https://images.unsplash.com/photo-1560472354-b33ff0c44a43?ixlib=rb-4.0.3&auto=format&fit=crop&w=800&q=80",
+      attendees: "200+"
+    },
+    {
+      id: 3,
+      title: "International Investment Roundtable",
+      date: "September 20, 2025",
+      location: "Jeonju Chamber of Commerce",
+      featured: true,
+      category: "INVESTMENT",
+      image: "https://images.unsplash.com/photo-1552664730-d307ca884978?ixlib=rb-4.0.3&auto=format&fit=crop&w=800&q=80",
+      attendees: "150+"
+    },
+    {
+      id: 4,
+      title: "Workshop: GMP Compliance for Startups",
+      date: "November 5, 2025",
+      location: "JB SQUARE Incubation Center",
+      featured: false,
+      category: "WORKSHOP",
+      image: "https://images.unsplash.com/photo-1542744173-8e7e53415bb0?ixlib=rb-4.0.3&auto=format&fit=crop&w=800&q=80",
+      attendees: "50+"
+    },
+    {
+      id: 5,
+      title: "Agritech Demo Day 2025",
+      date: "November 12, 2025",
+      location: "Jeonbuk National University",
+      featured: false,
+      category: "DEMO DAY",
+      image: "https://images.unsplash.com/photo-1573496799652-408c2ac9fe98?ixlib=rb-4.0.3&auto=format&fit=crop&w=800&q=80",
+      attendees: "300+"
+    },
+    {
+      id: 6,
+      title: "Networking Night: Bio-Pharma Professionals",
+      date: "November 20, 2025",
+      location: "The Grand Hill Hotel, Jeonju",
+      featured: true,
+      category: "NETWORKING",
+      image: "https://images.unsplash.com/photo-1522202176988-66273c2fd55f?ixlib=rb-4.0.3&auto=format&fit=crop&w=800&q=80",
+      attendees: "250+"
+    },
+    {
+      id: 7,
+      title: "Webinar: The Future of CRISPR Technology",
+      date: "December 2, 2025",
+      location: "Online",
+      featured: false,
+      category: "WEBINAR",
+      image: "https://images.unsplash.com/photo-1587825140708-df876c12b44e?ixlib=rb-4.0.3&auto=format&fit=crop&w=800&q=80",
+      attendees: "1000+"
+    },
+    {
+      id: 8,
+      title: "JB Bio-Cluster Open House",
+      date: "December 10, 2025",
+      location: "JB SQUARE Main Campus",
+      featured: false,
+      category: "OPEN HOUSE",
+      image: "https://images.unsplash.com/photo-1582213782179-e0d53f98f2ca?ixlib=rb-4.0.3&auto=format&fit=crop&w=800&q=80",
+      attendees: "400+"
+    },
+    {
+      id: 9,
+      title: "End-of-Year Gala & Awards Ceremony",
+      date: "December 18, 2025",
+      location: "Jeonju City Hall",
+      featured: true,
+      category: "GALA",
+      image: "https://images.unsplash.com/photo-1527529482837-4698179dc6ce?ixlib=rb-4.0.3&auto=format&fit=crop&w=800&q=80",
+      attendees: "350+"
+    }
+  ];
+
+  return (
+    <section id={id} className="py-24 bg-gradient-to-br from-slate-50 via-purple-50 to-blue-50">
+      <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
+        <motion.div
+          initial={{ opacity: 0, y: 30 }}
+          whileInView={{ opacity: 1, y: 0 }}
+          viewport={{ once: true }}
+          className="text-center mb-16"
+        >
+          <h2 className="text-5xl font-black text-[#0B2D63] mb-6" dangerouslySetInnerHTML={{ __html: t('events.title') }}>
+          </h2>
+          <p className="text-xl text-slate-600 max-w-3xl mx-auto font-medium">
+            {t('events.subtitle')}
+          </p>
+        </motion.div>
+
+        <motion.div
+          variants={staggerBold}
+          initial="initial"
+          whileInView="animate"
+          viewport={{ once: true }}
+          className="grid md:grid-cols-2 lg:grid-cols-3 gap-8"
+        >
+          {events.map((event) => (
+            <motion.div
+              key={event.id}
+              variants={boldFadeIn}
+              whileHover={{ scale: 1.05, y: -10 }}
+              className={`bg-white rounded-2xl overflow-hidden shadow-xl hover:shadow-2xl transition-all cursor-pointer ${
+                event.featured ? 'ring-4 ring-[#FFC940] ring-opacity-50' : 'border-2 border-slate-200'
+              }`}
+            >
+              <div className="h-56 relative overflow-hidden">
+                <img
+                  src={event.image}
+                  alt={event.title}
+                  className="w-full h-full object-cover transform hover:scale-110 transition-transform duration-700"
+                />
+                <div className="absolute inset-0 bg-gradient-to-t from-black/80 via-black/20 to-transparent"></div>
+
+                <div className="absolute top-4 left-4 right-4 flex justify-between items-start">
+                  {event.featured && (
+                    <span className="px-4 py-2 bg-[#FFC940] text-[#0B2D63] text-xs font-black rounded-full uppercase tracking-wider shadow-lg">
+                      🔥 {t('events.featured')}
+                    </span>
+                  )}
+                  <span className="px-3 py-1 bg-white/20 backdrop-blur-md text-white text-xs font-bold rounded-full border border-white/30">
+                    {t('events.attendees', { count: event.attendees })}
+                  </span>
+                </div>
+
+                <div className="absolute bottom-4 left-4 right-4">
+                  <span className="px-3 py-1 text-xs font-bold rounded-full bg-white/20 backdrop-blur-md text-white border border-white/30 uppercase tracking-wider">
+                    {event.category}
+                  </span>
+                </div>
+              </div>
+
+              <div className="p-6">
+                <div className="flex items-center text-sm text-slate-500 mb-3 font-semibold">
+                  <Calendar className="w-4 h-4 mr-2" />
+                  {event.date}
+                </div>
+                <h3 className="text-xl font-black text-[#0B2D63] mb-4 leading-tight">{event.title}</h3>
+                <div className="flex items-center text-sm text-slate-600 mb-6 font-medium">
+                  <MapPin className="w-4 h-4 mr-2" />
+                  {event.location}
+                </div>
+                <button className={`w-full px-6 py-3 font-black rounded-xl transition-all transform hover:scale-105 ${
+                  event.featured
+                    ? 'bg-[#FFC940] text-[#0B2D63] hover:bg-[#FFB800] shadow-lg'
+                    : 'bg-[#0B2D63] text-white hover:bg-[#1a4480] shadow-lg'
+                }`}>
+                  {t('events.register')}
+                </button>
+              </div>
+            </motion.div>
+          ))}
+        </motion.div>
+      </div>
+    </section>
+  );
+};
+
+export default EventsSection;
