+++ src/components/types.ts
@@ -1,0 +1,3 @@
+export interface SectionProps {
+  id: string;
+}
